Updated SSL auth
According to the function 'perform_ssl_authentication()' inside of woocommerce/includes/api/class-wc-api-authentication.php, the consumer_key and consumer_secret must be passed in the URL as GET parameters when authenticating over SSL.

Without the change I'm proposing, I was unable to get this module to work with my SSL-enabled server. The error message was consistently "Consumer Secret is invalid".

This change is working smoothly for me in a production environment, however if this change is not needed and I'm doing something wrong I would appreciate the feedback.