api accepts url params and additional keyword args for requests module
Currently the url params that are sent to Woo-Commerce are a closed box.
They should be exposed outside of the function. Additionally, there are
cases where a developer needs to pass additional parameters to the
requests library. These should also be exposed outside of the function.

This pull request simply achieves these things.